// src/types/product-field.ts
export type ProviderInfoField =
    | 'type_transaction_id'
    | 'status_transaction_sell_id'
    | 'status_transaction_rent_id'
    | 'supplier_type_id'
    | 'send_date'
    | 'customer_id'
    | 'ref_id'
    | 'expert_id'
    | 'supplier_note'
    | 'product_code'

export type RealEstateInfoField =
    | 'name'
    | 'slug'
    | 'address'
    | 'type_product_id'
    | 'zone_province_id'
    | 'zone_ward_id'
    | 'project_id'
    | 'identifier_code'
    | 'product_code'
    | 'publish_status'

export type PriceInfoField =
    | 'status_product_id'
    | 'acreage'
    | 'price_build'
    | 'price_contact'
    | 'total_contact'
    | 'difference'
    | 'percent_closed'
    | 'total_closed'
    | 'price_receive'
    | 'total_price_receive'
    | 'price_sell'
    | 'total_price_sell'
    | 'percent_brokerage'
    | 'price_brokerage'
    | 'tax_type_id'
    | 'price_rent'
    | 'deposit_rent'
    | 'cycle_rent'
    | 'price_rent_brokerage'
    | 'price_note'
    | 'product_value'

export type ApartmentDetailField =
    | 'apartment'
    | 'number_floor'
    | 'number'
    | 'is_corner'
    | 'finish_house'
    | 'direction_house_id'
    | 'direction_balcony_id'
    | 'number_bedrooms'
    | 'number_toilets'
    | 'description'

export type KioskDetailField =
    | 'apartment'
    | 'number_floor'
    | 'number'
    | 'is_corner'
    | 'finish_house'
    | 'direction_house_id'
    | 'street_frontage'
    | 'description'

export type LandDetailField =
    | 'address'
    | 'number'
    | 'finish_house'
    | 'direction_house_id'
    | 'street_frontage'
    | 'gateway'
    | 'description'

export type PrivateHouseDetailField =
    | 'address'
    | 'number'
    | 'finish_house'
    | 'direction_house_id'
    | 'direction_balcony_id'
    | 'street_frontage'
    | 'gateway'
    | 'number_bedrooms'
    | 'number_toilets'
    | 'description'

export type ResortDetailField = 'address' | 'finish_house' | 'description'

export type TownHouseDetailField =
    | 'divisive'
    | 'parcel'
    | 'number'
    | 'is_corner'
    | 'finish_house'
    | 'direction_house_id'
    | 'street_frontage'
    | 'gateway'
    | 'number_bedrooms'
    | 'number_toilets'
    | 'description'